core_modules/requirements.txt
numpy==1.21.6
pathlib
Pillow
opencv-python